[0.3.5]:"更新了多个接口以及文件上传功能的优化" [ 1.File  [0.3.5 NEW] 2.ActivityAudit [0.3.5 NEW] 3.MentorAudit [0.3.5 NEW] 4.StudentAudit [0.3.5 NEW] 5.CreditsAudit [0.3.5 NEW] 6./project/projectSearchFuzzy [0.3.5 NEW] 7./project/projectSearch [0.3.5 UPDATE] 8./mentor/projectAdd [0.3.5 NEW] 9./mentor/projectManagementAdd [0.3.5弃用] 10.新增 student/mentor/credits 的 Audits 以及 mail project_Forum 的相关数据表 ]

+++ src/main/java/com/example/university_project_platform_backend/entity/ProjectManagement.java
@@ -13,7 +13,7 @@ import lombok.Setter;
  * </p>
  *
  * @author blackhaird
- * @since 2024-04-15
+ * @since 2024-04-19
  */
 @Getter
 @Setter
@@ -48,7 +48,7 @@ public class ProjectManagement implements Serializable {
     /**
      * 项目状态id 0代表未通过 1代表通过 2代表审核中 
      */
-    private Integer projectStatusId;
+    private Byte projectStatusId;
 
     /**
      * 项目状态状态描述 注释/备注
